Bot/requirements.txt
discord.py>=2.3.2
scikit-learn>=1.3.0
requests>=2.31.0
pydub==0.25.1
python-dotenv>=1.0.0
numpy==1.24.3
psutil==5.9.5
pygame==2.5.0
beautifulsoup4>=4.12.2
PyNaCl==1.5.0
ffmpeg-python>=0.2.0 